file renamed to video_recording
