+++ tools/sense_studio/video_recording.py
